+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- CurveOperator: FindClosestPoint
+- NurbsCurve: FindClosestPoint and FindClosestPointWithInitialGuess
+- NurbsSurface: FindClosestPointWithInitialGuess
 
 ### Changed